+++ pkg/query/flamegraph_arrow.go
@@ -89,6 +89,7 @@ func GenerateFlamegraphArrow(
 	w := ipc.NewWriter(&buf,
 		ipc.WithSchema(record.Schema()),
 		ipc.WithAllocator(mem),
+		ipc.WithLZ4(),
 	)
 	defer w.Close()