+++ ui/packages/shared/profile/src/ProfileFlameGraph/index.tsx
@@ -77,7 +77,6 @@ export const validateFlameChartQuery = (
 ): {isValid: boolean; isNonDelta: boolean; isDurationTooLong: boolean} => {
   const isNonDelta = !profileSource.ProfileType().delta;
   const duration = profileSource.mergeTo - profileSource.mergeFrom;
-  console.log('duration of flame chart query: ', duration, 'ns');
   const isDurationTooLong = duration > 60_000_000_000n; // 60 seconds in nanoseconds
   return {isValid: !isNonDelta && !isDurationTooLong, isNonDelta, isDurationTooLong};
 };