+++ ui/packages/shared/profile/src/ProfileView/components/Toolbars/index.tsx
@@ -47,6 +47,8 @@ export interface VisualisationToolbarProps {
   preferencesModal?: boolean;
   profileViewExternalSubActions?: React.ReactNode;
   setGroupByLabels: (labels: string[]) => void;
+  flamechartDimension: string[];
+  setFlamechartDimension: (labels: string[]) => void;
   showVisualizationSelector?: boolean;
   sandwichFunctionName?: string;
   alignFunctionName: string;
@@ -135,6 +137,8 @@ export const VisualisationToolbar: FC<VisualisationToolbarProps> = ({
   groupBy,
   toggleGroupBy,
   setGroupByLabels,
+  flamechartDimension,
+  setFlamechartDimension,
   profileType,
   profileSource,
   queryClient,
@@ -158,6 +162,8 @@ export const VisualisationToolbar: FC<VisualisationToolbarProps> = ({
   const isTableVizOnly = dashboardItems?.length === 1 && isTableViz;
   const isGraphViz = dashboardItems?.includes('flamegraph');
   const isGraphVizOnly = dashboardItems?.length === 1 && isGraphViz;
+  const isFlamechartViz = dashboardItems?.includes('flamechart');
+  const isFlamechartVizOnly = dashboardItems?.length === 1 && isFlamechartViz;
 
   const req = profileSource?.QueryRequest();
   if (req !== null && req !== undefined) {
@@ -184,8 +190,19 @@ export const VisualisationToolbar: FC<VisualisationToolbarProps> = ({
             </>
           )}
 
+          {isFlamechartViz && (
+            <GroupByDropdown
+              groupBy={flamechartDimension}
+              labels={groupByLabels}
+              setGroupByLabels={setFlamechartDimension}
+              metadataRefetch={metadataRefetch}
+              metadataLoading={metadataLoading}
+              label="Samples group by"
+            />
+          )}
+
           <div className="flex mt-5">
-            <ProfileFilters />
+            {!isFlamechartVizOnly && <ProfileFilters />}
 
             {profileViewExternalSubActions != null ? profileViewExternalSubActions : null}
           </div>