+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.1004](https://github.com/parca-dev/parca/compare/@parca/web@0.16.1003...@parca/web@0.16.1004) (2026-02-11)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.1003](https://github.com/parca-dev/parca/compare/@parca/web@0.16.1002...@parca/web@0.16.1003) (2026-02-10)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.1002](https://github.com/parca-dev/parca/compare/@parca/web@0.16.1001...@parca/web@0.16.1002) (2026-02-09)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.1001](https://github.com/parca-dev/parca/compare/@parca/web@0.16.1000...@parca/web@0.16.1001) (2026-02-09)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.1000](https://github.com/parca-dev/parca/compare/@parca/web@0.16.999...@parca/web@0.16.1000) (2026-02-06)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.999](https://github.com/parca-dev/parca/compare/@parca/web@0.16.998...@parca/web@0.16.999) (2026-01-29)
 
 **Note:** Version bump only for package @parca/web