+++ ui/packages/shared/store/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.200 (2026-02-06)
+
+**Note:** Version bump only for package @parca/store
+
 ## [0.16.199](https://github.com/parca-dev/parca/compare/@parca/store@0.16.198...@parca/store@0.16.199) (2025-12-15)
 
 **Note:** Version bump only for package @parca/store